move database to database_handler
