--- database/src/database/mod.rs
@@ -1,3 +1,0 @@
-mod database;
-
-pub use database::Database;
