+++ database/src/database_handler/mod.rs
@@ -1,0 +1,1 @@
+pub mod database_handler;
