+++ database/src/lib.rs
@@ -1,6 +1,6 @@
 pub mod btree;
 pub mod command;
-pub mod database;
+pub mod database_handler;
 pub mod protocol;
 pub mod server;
 pub mod storage;
